+++ src/app/app.routes.ts
@@ -1,4 +1,15 @@
-import { Routes } from '@angular/router';
+import { inject } from '@angular/core';
+import { ResolveFn, Routes } from '@angular/router';
+import { lastValueFrom } from 'rxjs';
+import {
+  BlogBackendService,
+  Entries,
+} from './core/services/blog-backend.service';
+
+export const entriesResolver: ResolveFn<Entries> = async () => {
+  const blogBackendService = inject(BlogBackendService);
+  return await lastValueFrom(blogBackendService.getBlogPosts());
+};
 
 export const APP_ROUTES: Routes = [
   {
@@ -10,5 +21,6 @@ export const APP_ROUTES: Routes = [
     path: 'overview',
     loadChildren: () =>
       import('./features/blog-overview-page/blog-overview-page.routes'),
+    resolve: { model: entriesResolver },
   },
 ];
