+++ src/app/core/services/blog-backend.service.ts
@@ -1,0 +1,44 @@
+import { inject, Injectable } from '@angular/core';
+import { map, Observable } from 'rxjs';
+import { environment } from '../../../enviroments/environment';
+import { HttpClient } from '@angular/common/http';
+import { z } from 'zod';
+
+const BlogSchema = z.object({
+  id: z.number(),
+  title: z.string(),
+  contentPreview: z.string(),
+  author: z.string(),
+  likes: z.number(),
+  comments: z.number(),
+  likedByMe: z.boolean(),
+  createdByMe: z.boolean(),
+  headerImageUrl: z.string().optional(),
+});
+
+const BlogArraySchema = z.array(BlogSchema);
+
+const EntriesSchema = z.object({
+  data: BlogArraySchema,
+  pageIndex: z.number(),
+  pageSize: z.number(),
+  totalCount: z.number(),
+  maxPageSize: z.number(),
+});
+
+export type Blog = z.infer<typeof BlogSchema>;
+
+export type Entries = z.infer<typeof EntriesSchema>;
+
+@Injectable({
+  providedIn: 'root',
+})
+export class BlogBackendService {
+  httpClient = inject(HttpClient);
+
+  getBlogPosts(): Observable<Entries> {
+    return this.httpClient
+      .get<Entries>(`${environment.serviceUrl}/entries`)
+      .pipe(map((entries) => EntriesSchema.parse(entries)));
+  }
+}
