--- src/app/features/blog-overview-page/blog-overview-page.component.spec.ts
@@ -1,22 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { BlogOverviewPageComponent } from './blog-overview-page.component';
-
-describe('BlogOverviewPageComponent', () => {
-  let component: BlogOverviewPageComponent;
-  let fixture: ComponentFixture<BlogOverviewPageComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [BlogOverviewPageComponent],
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(BlogOverviewPageComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
