+++ src/app/features/blog-overview-page/blog-overview-page.component.ts
@@ -1,38 +1,9 @@
 import { NgIf, AsyncPipe } from '@angular/common';
-import { HttpClient } from '@angular/common/http';
-import { Component, inject } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
 import { MatIcon } from '@angular/material/icon';
-import { Observable, map } from 'rxjs';
-import { environment } from '../../../enviroments/environment';
-import { z } from 'zod';
-
-const BlogSchema = z.object({
-  id: z.number(),
-  title: z.string(),
-  contentPreview: z.string(),
-  author: z.string(),
-  likes: z.number(),
-  comments: z.number(),
-  likedByMe: z.boolean(),
-  createdByMe: z.boolean(),
-  headerImageUrl: z.string().optional(),
-});
-
-const BlogArraySchema = z.array(BlogSchema);
-
-const EntriesSchema = z.object({
-  data: BlogArraySchema,
-  pageIndex: z.number(),
-  pageSize: z.number(),
-  totalCount: z.number(),
-  maxPageSize: z.number(),
-});
-
-export type Blog = z.infer<typeof BlogSchema>;
-
-export type Entries = z.infer<typeof EntriesSchema>;
+import { Entries } from '../../core/services/blog-backend.service';
 
 @Component({
   selector: 'app-blog-overview-page',
@@ -42,12 +13,5 @@ export type Entries = z.infer<typeof EntriesSchema>;
   styleUrl: './blog-overview-page.component.scss',
 })
 export class BlogOverviewPageComponent {
-  httpClient = inject(HttpClient);
-  blogs$: Observable<Entries>;
-
-  constructor() {
-    this.blogs$ = this.httpClient
-      .get<Entries>(`${environment.serviceUrl}/entries`)
-      .pipe(map((entries) => EntriesSchema.parse(entries)));
-  }
+  @Input({ required: true }) model!: Entries;
 }
